pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-ducklake"
version = "0.0.1"
description = "DuckLake provider for Apache Airflow (based on DuckDB)"
authors = [{ name = "Your Name", email = "schapman1974@gmail.com" }]
license = { text = "Apache License 2.0" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["apache-airflow>=2.0", "duckdb>=0.5.0"]
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
urls = { Homepage = "https://github.com/schapman1974/airflow-provider-ducklake" }

[project.entry-points."apache_airflow_provider"]
provider_info = "ducklake_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
readme = { file = "README.md", content-type = "text/markdown" }

[tool.setuptools.packages.find]
where = ["."]
namespaces = false
